refactored get_mx into lib/get_mx
