+++ lib/HarakaMx.js
@@ -7,7 +7,7 @@ class HarakaMx {
   constructor(obj = {}, domain) {
     switch (typeof obj) {
       case 'string':
-        /mtp:\/\//.test(obj) ? this.fromUrl(obj) : this.fromString(obj)
+        ;/mtp:\/\//.test(obj) ? this.fromUrl(obj) : this.fromString(obj)
         break
       case 'object':
         this.fromObject(obj)
@@ -82,4 +82,4 @@ class HarakaMx {
   }
 }
 
-exports.HarakaMx = HarakaMx
+module.exports = HarakaMx
