+++ package.json
@@ -4,7 +4,8 @@
   "description": "haraka network utilities",
   "main": "index.js",
   "files": [
-    "CHANGELOG.md", "lib"
+    "CHANGELOG.md",
+    "lib"
   ],
   "scripts": {
     "format": "npm run prettier:fix && npm run lint:fix",
