+++ test/get_ip_any.js
@@ -1,7 +1,6 @@
 const assert = require('node:assert')
 const net = require('node:net')
 
-
 const ip_fixtures = [
   [false, ' 2001:0000:1234:0000:0000:C1C0:ABCD:0876  '],
   [false, ' 2001:0000:1234:0000:0000:C1C0:ABCD:0876  0'],
